% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{bgSub}
\alias{bgSub}
\title{Subtract a background spectrum}
\usage{
bgSub(spectra, bgSpectrum)
}
\arguments{
\item{spectra}{Matrix containing spectral data, typically returned by \code{openSpectraFile} or \code{getSpectra}.}

\item{bgSpectrum}{Numeric vector where \code{length(bgSpectrum) == nrow(spectra)}.}
}
\value{
Matrix of same dimensions as \code{spectra} containing background-subtracted spectra.
}
\description{
Given a matrix containing spectral data and a vector containing a background spectrum, subtract the background spectrum from each spectrum in the matrix.
}
\examples{
blanks <- getSpectra(TuSC_Spc110_spectra, nRows=9, nCols=3, startRow=1, endRow=2,
startColumn=1, endColumn=3)
spectra <- getSpectra(TuSC_Spc110_Spectra, nRows=9, nCols=3, startRow=3, endRow=9,
startColumn=1, endColumn=3)
spectra_bgsub <- bgSub(spectra=spectra, bgSpectrum=rowMeans(blanks))

}
