% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfp_yfp_ref}
\alias{cfp_yfp_ref}
\title{Basis spectra for linear unmixing of CFP/YFP spectra}
\format{A matrix with 29 rows and 2 columns. The \code{dimnames} attributes
indicate CFP or YFP and the wavelengths at which the spectra were recorded 
in the columns and rows, respectively.}
\source{
Lyon et al. 2016. Molecular Biology of the Cell 27: 2245, figure S1A. \url{http://www.molbiolcell.org/content/27/14/2245.long}
}
\usage{
cfp_yfp_ref
}
\description{
Fluorescence spectra from gamma-tubulin small complex (gTuSC) containing 
Spc97-YFP or Spc98-CFP were recorded on a Molecular Devices SpectraMax
M5 plate reader. For gTuSC-CFP, the excitation wavelength was 420 nm with
emission recorded through a 455 nm longpass filter in 5 nm increments from 460
to 600 nm. Spectra were recorded in the same manner for gTuSC-YFP, but with
excitation at 475 nm, a 495 nm longpass filter, and spectra recorded from 495
to 600 nm. YFP emission from 460 to 490 nm is set to zero. Multiple spectra
were recorded, background subtracted, and averaged. The spectra were then 
scaled so that the maximum intensity is one.
}
\keyword{datasets}
