% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBinding.R
\name{fitBinding}
\alias{fitBinding}
\title{Fit a binding model to data}
\usage{
fitBinding(data, kdGuess = "auto", probeConcentration, hillCoefficient = 2,
  model = c("s", "t", "c"), weight = TRUE, xCol = 1, yCol = 2,
  uncCol = 3)
}
\arguments{
\item{data}{Dataframe with at least two columns, typically the output of \code{fretFit}. Columns must contain an independent variable, observed values. An optional column may contain uncertainties for the observed values.}

\item{kdGuess}{Either \code{"auto"} or numeric. If \code{"auto"}, an initial estimate for the dissociation constant is determined from the data. Alternatively, users may specify an initial estimate as a numeric value.}

\item{probeConcentration}{Numeric. For \code{model="t"}, the concentration of the non-titrated component in the binding experiment in the same units as the independent variable in \code{data}.}

\item{hillCoefficient}{Numeric. Initial estimate of the Hill coefficient for \code{model="c"}.}

\item{model}{One of \code{"s"}, \code{"t"}, or \code{"c"}. The binding model to fit: \code{"s"} for simple, \code{"t"} for tight, \code{"c"} for cooperative.}

\item{weight}{Logical. Perform weighted non-linear least squares fit using user-provided uncertainties?}

\item{xCol}{Integer. The column in \code{data} containing the independent variable.}

\item{yCol}{Integer. The column in \code{data} containing observed values.}

\item{uncCol}{Integer. The column in \code{data} containing uncertainties for the observed values. Required if \code{weight=TRUE}.}
}
\value{
Object of class \code{"nls"} with optimized binding model parameters.
}
\description{
Given a dataframe with an independent variable (typically concentration) and observed values (typically FRET or fluorescence intensity) with optional uncertanties, this function fits one of three binding models using the Levenberg-Marquardt non-linear least squares optimizer. The three available models are
\itemize{
    \item \code{"s"} for simple, single-site saturable binding
    \item \code{"t"}  for tight single-site saturable binding where total and unbound concentrations of the titrated
          component cannot be assumed equal (i.e. when the non-titrated component concentration is similar to the
           dissociation constant)
    \item \code{"c"} for cooperative binding
}
}
\details{
Binding model definitions are as follows, where \eqn{B_{max}} is the theoretical observed value as the independent variable approaches infinity, \eqn{B_{min}} is the observed value when the independent variable is zero, \eqn{K_{d}} is the equilibrium dissociation constant, \eqn{c} is the concentration of the non-titrated component, and \eqn{n} is the Hill coefficient.

Simple:
\deqn{y = (B_{max}-B_{min})\frac{x}{x + K_d}+B_{min}}

Tight:
\deqn{y= (B_{max}-B_{min})\frac{(x + c + K_d) - \sqrt{(x + c + K_d)^2 - 4xc}}{2c}+B_{min}}

Cooperative:
\deqn{y = (B_{max}-B_{min})\frac{x^n}{x^n + K_d^n}+B_{min}}

The \code{kdGuess = "auto"} option finds two data points close to the half-maximal observed value, fits a line to them, then interpolates to find the x-axis value corresponding to the half-maximum point on the binding curve. It is likely to have problems with noisy data, data with outliers, or binding curves where saturation has not been reached.

Weighted least-squares is performed with weights \eqn{1/\sigma^2}, so the user should supply uncertainties as standard deviations.
}
\examples{
fitBinding(data=TuSC_Spc110_binding, model="s")
}
