% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretFit.R
\name{fretFit}
\alias{fretFit}
\title{Fit spectra as a linear combination of basis spectra}
\usage{
fretFit(spectra, concentrations, accConc, accCorrect = TRUE,
  normalize = TRUE, fitted = FALSE, average = TRUE, basis = cfp_yfp_ref,
  donorCol = 1, accCol = 2, accCorr = c(slope = 0.3986, intercept =
  -0.7486))
}
\arguments{
\item{spectra}{Matrix with spectral data arranged so that each column contains one spectrum.}

\item{concentrations}{Numeric vector of concentrations, for instance the concentrations of a titrated component in a binding experiment. The order of concentrations must correspond to the ordering of columns in the \code{spectra} matrix. Required if \code{fitted = FALSE}.}

\item{accConc}{Numeric. The concentration of the acceptor fluorophore. Required if \code{accCorrect = TRUE} to allow for correction for direct excitation of the acceptor fluorophore in FRET experiments.}

\item{accCorrect}{Logical. If \code{TRUE}, the contribution of the acceptor signal due to direct excitation (i.e., not FRET) is subtracted before further analysis.}

\item{normalize}{Logical. If \code{TRUE} FRET is calculated as Acceptor/(Donor + Acceptor). Otherwise FRET is simply Acceptor/Donor.}

\item{fitted}{Logical. If \code{TRUE}, this function returns a list with components \code{Fitted}, a matrix with fitted spectral values, and \code{Coefs}, containing the coefficients from the least-squares fit.}

\item{average}{Logical. If \code{TRUE} calculates mean FRET values and standard deviations. Otherwise unaveraged FRET values are returned.}

\item{basis}{Matrix where \code{nrow(basis) == nrow{spectra}}. Basis spectra for the least-squares fit. The default \code{cfp_yfp_ref} is basis spectra for CFP-YFP FRET experiments. See \code{?cfp_yfp_ref}.}

\item{donorCol}{Numeric integer. The column of \code{basis} corresponding to the donor fluorophore in FRET experiments.}

\item{accCol}{Numeric integer. The column of \code{basis} corresponding to the acceptor fluorophore in FRET experiments.}

\item{accCorr}{Numeric vector with components \code{slope} and \code{intercept}. Values for linear correction term for subtracting acceptor fluorophore signal due to direct excitation of the acceptor (i.e., not FRET). See details.}
}
\value{
A dataframe with FRET values, unless \code{fitted=TRUE} in which a list with components \code{Fitted}, a matrix with fitted spectral values, and \code{Coefs}, containing the coefficients from the least-squares fit.
}
\description{
A spectrum recorded from a sample containing more than one fluorophore (or chromophore) is the linear combination of the fluorophores in isolation. Spectra from samples containing unknown contributions from multiple fluorophores can be "unmixed" by least-squares fitting as the sum of multiple basis spectra. This is particularly useful for Forster resonance energy transfer (FRET) experiments. This function performs the least-squares fit and provides several tools useful in FRET applications.
}
\examples{
#Get the background spectrum
background_spectrum <- rowMeans(getSpectra(TuSC_Spc110_spectra, nRows=9, nCols=3, startRow=1,
endRow=2, startColumn=1, endColumn=3))
#Subtract background from experimental spectra
TuSC_Spc110_spectra_bgsub <- bgSub(getSpectra(TuSC_Spc110_spectra, nRows=9, nCols=3,
startRow=3, endRow=9, startColumn=1, endColumn=3), background_spectrum)
#Fit the spectra and get FRET values with statistics
TuSC_Spc110_binding <- fretFit(TuSC_Spc110_spectra_bgsub, concentrations=sort(rep(c(1500,
750, 375, 187.5, 93.75, 46.875, 0), 3)), accConc=50, average=TRUE)
#Fit the spectra and get the fitted spectra
TuSC_Spc110_binding <- fretFit(TuSC_Spc110_spectra_bgsub, concentrations=sort(rep(c(1500,
750, 375, 187.5, 93.75, 46.875, 0), 3)), accConc=50, fitted=TRUE)
#Fit the spectra and get unaveraged FRET values
TuSC_Spc110_binding <- fretFit(TuSC_Spc110_spectra_bgsub, concentrations=sort(rep(c(1500,
750, 375, 187.5, 93.75, 46.875, 0), 3)), accConc=50, average=FALSE)
}
