% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{getSpectra}
\alias{getSpectra}
\title{Get a set of spectra corresponding to a group of wells from a microplate}
\usage{
getSpectra(spectra, nRows, nCols, startRow, endRow = startRow, startColumn,
  endColumn = startColumn)
}
\arguments{
\item{spectra}{Matrix, typically returned by \code{openSpectraFile}. See details for required features of this matrix.}

\item{nRows}{Numeric or length 2 character vector. If numberic, the number of rows containing samples in the microplate. If character vector, the letter codes for the first and last rows containing sample.}

\item{nCols}{Numeric. The number of columns containing samples in the microplate.}

\item{startRow}{Numeric or character. The first row containing samples of interest. Can be specified as the ordinal number for the row of interest or the letter code.}

\item{endRow}{Numeric or character. The last row containing samples of interest. Can be specified as the ordinal number for the row of interest or the letter code.}

\item{startColumn}{Numeric. The first column containing samples of interest.}

\item{endRow}{Numeric. The last column containing samples of interest.}
}
\value{
A matrix where each column is a spectrum.
}
\description{
Microplate experiments often have a set of control samples and a set of experimental 
samples (for example) where it might be convenient to separate the spectra into 
different objects for processing and analysis. Given an input matrix containing 
spectral data from a microplate experiment, this function returns the spectra from 
a rectangular block of wells.
}
\details{
The \code{spectra} matrix must be arranged so that the groups of adjacent columns correspond to wells within a row of a microplate. That is, for a microplate with rows labeled A, B, C,  etc., and columns labeled 1, 2, 3, 4, etc., the columns of the \code{spectra} matrix correspond to wells A1, A2, A3, A4, B1, B2, B3, B4, C1, C2, C3, C4, etc., in that order.
}
\examples{
#The matrix TuSC_110_spectra was recorded in wells A7-I9 of a microplate,
with buffer-only controls in wells A7-B9.
#Get the spectra for the buffer only controls
getSpectra(TuSC_110_spectra, nRows=9, nCols=3, startRow=1, endRow=2, startColumn=1,
endColumn=3)
#Get the experimental spectra using row numbers
getSpectra(TuSC_Spc110_spectra, nRows=9, nCols=3, startRow=3, endRow=9, startColumn=1,
endColumn=3)
#Get the experimental spectra using row letters
getSpectra(TuSC_Spc110_spectra, nRows=c("A", "I"), ncols=3, startRow="C", endRow="I",
startColumn=1, endColumn=3)

}
