% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{openSpectraFile}
\alias{openSpectraFile}
\title{Open a file containing spectral data}
\usage{
openSpectraFile(file = "")
}
\arguments{
\item{file}{Character string giving absolute or relative path to a tab-delimited data file.}
}
\value{
A matrix where each column is a spectrum.
}
\description{
Simple wrapper around \code{read.table} for opening tab-delimited spectral data files.
}
