% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinding.R
\name{plotBinding}
\alias{plotBinding}
\title{Plot binding data}
\usage{
plotBinding(data, xlab = "Concentration (nM)", ylab = "FRET", ylim = NULL,
  xlim = NULL, pch = 21, cex = 1, xlog = F, logBase = 10,
  col = "black", bg = 1, errCol = "black", errLwd = 1,
  errFeet = FALSE, main = NULL, xCol = 1, yCol = 2, uncCol = 3, ...)
}
\arguments{
\item{data}{Dataframe with at least two columns, typically returned by \code{fretFit}. The two required columns must contain independent variable and one observations. Uncertainties in the observed values may be included as an optional third column.}

\item{xlab}{Character. X-axis label.}

\item{ylab}{Character. Y-axis label.}

\item{ylim}{Length two numeric vector. The lower and upper boundaries of the vertical axis of the plot.}

\item{xlim}{Length two numeric vector. The left and right boundaries of the horizontal axis of the plot.}

\item{pch}{Any valid plot character (numeric or character).}

\item{cex}{Plot character magnification relative to default size (e.g. \code{cex=2} will make the plot character two times larger than default).}

\item{xlog}{Logical. Should the x-axis values be logarithmically transformed?}

\item{logBase}{Numeric. If \code{xlog=TRUE}, the base of the logarithmic transform of the x-axis values.}

\item{col}{Any valid color specification. The foreground color of the plot characters.}

\item{bg}{Any valid color specification. The background color of the plot characters.}

\item{errCol}{Any valid color specification. The color of the error bars.}

\item{errLwd}{Numeric. The line weight of the error bars. \code{errLwd = 1} is a line 1/96 inch thick, \code{errLwd = 0.75} is a line 1 point thick.}

\item{errFeet}{Logical. Should horizontal lines ("feet") be drawn on the error bars? Default \code{FALSE} will draw vertical lines only.}

\item{main}{Character. Title for plot.}

\item{xCol}{Integer. The column of \code{data} containing the independent variable.}

\item{yCol}{Integer. The column of \code{data} containing observed values.}

\item{uncCol}{Integer. Optional. The column of \code{data} containing uncertainties in observed values.}

\item{...}{Additional arguments to \code{plot} or \code{points}}
}
\value{
No return value, changes state of graphics device.
}
\description{
Plot binding data as scatter plot with error bars.
}
\examples{
#Plot with linear x-axis scale
plotBinding(TuSC_Spc110_binding)
#Plot with "feet" on the error bars the same color as the plot symbols
plotBinding(TuSC_Spc110_binding, errFeet=TRUE, pch=19, col='red', errCol='red')
#Plot with log-10 x-axis scale
plotBinding(TuSC_Spc110_binding, xlog=T)
}
