% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot a binding model}
\usage{
plotFit(model, xRange = NULL, xlog = FALSE, logBase = 10, add = TRUE,
  col = 1, lwd = 1, lty = 1, xlab = "Concentration (nM)",
  ylab = "FRET", xlim = NULL, ylim = NULL, main = NULL, ...)
}
\arguments{
\item{model}{The model object for which to plot a curve. Typically returned by \code{fitBinding}, though models returned by \code{lm}, \code{nls} are likely to work as well as long as the fitting function was run with \code{model=TRUE}.}

\item{xRange}{Length-two numeric vector. The x-axis values between which the fit model will be evaluated and a curve plotted.}

\item{xlog}{Logical. Should the curve be drawn with a logarithmic x-axis scale?}

\item{logBase}{Numeric. The base of the logarithmic transform used if \code{xlog=TRUE}.}

\item{add}{Logical. Should the curve be added to an existing graphics device? If \code{FALSE}, the curve is drawn in a new graphics device.}

\item{col}{Any valid R color specification. The color of the curve.}

\item{lwd}{Numeric. The line width. \code{lwd=1} is 1/96 of an inch. \code{lwd=0.75} is one point.}

\item{lty}{Integer. The type of line to draw. 1 is solid, 2 is dashed, 3 is dotted. See \code{lty} under \code{?par} for more details.}

\item{xlab.}{Character. The x-axis label.}

\item{ylab.}{Character. The y-axis label.}

\item{xlim.}{Length-two numeric vector. The x-axis limits of the plot window for \code{add=FALSE}.}

\item{ylim.}{Length-two numeric vector. The y-axis limits of the plot window for \code{add=FALSE}.}

\item{main.}{Character. The title of the plot.}
}
\value{
No return value, changes state of graphics device.
}
\description{
Given a fit model (typically returned by \code{fitBinding}), plot a curve showing the fit evaluated over a range of x-axis values.
}
\examples{
#Fit a model to some data
fit <- fitBinding(TuSC_Spc110_binding)
#Plot the binding data
plotBinding(TuSC_Spc110_binding)
#Plot the model curve
plotFit(fit)
}
