% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateReader.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Plot a set of spectra}
\usage{
plotSpectra(spectra, wl = seq(460, 600, by = 5), xlab = "Wavelength (nm)",
  ylab = "Fluorescence intensity (AU)", bg = "grey80", ...)
}
\arguments{
\item{spectra}{Matrix containing spectral data, typically returned by \code{openSpectraFile} or \code{getSpectra}.}

\item{wl}{Numeric vector of wavelengths.}

\item{xlab}{Character. X-axis label.}

\item{ylab}{Character. Y-axis label.}

\item{bg}{Any valid R color specification. Background color of plot.}

\item{...}{Further arguments to \code{plot}.}
}
\value{
No return value, changes state of graphics device.
}
\description{
Given a matrix containing spectral data, plot each spectrum on the same set of axes.
}
\examples{
blanks <- getSpectra(TuSC_Spc110_spectra, nRows=9, nCols=3, startRow=1, endRow=2,
startColumn=1, endColumn=3)
plotSpectra(blanks)

}
