% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TuSC_Spc110_binding}
\alias{TuSC_Spc110_binding}
\title{Spc110-induced gTuRC assembly curve}
\format{7 x 4 dataframe with columns Concentration, FRET, SD, and N.}
\source{
Lyon et al. 2016. Molecular Biology of the Cell 27: 2245, figure 2E (Dimer WT curve). \url{http://www.molbiolcell.org/content/27/14/2245.long}
}
\usage{
TuSC_Spc110_binding
}
\description{
FRET data for gTuSC-CFP/YFP in the presence of varying concentrations of Spc110.
}
\examples{
#TuSC_Spc110_binding was prepared as follows:
background_spectrum <- rowMeans(getSpectra(TuSC_Spc110_spectra, nRows=9, nCols=3, startRow=1,
endRow=2, startColumn=1, endColumn=3))
TuSC_Spc110_spectra_bgsub <- bgSub(getSpectra(TuSC_Spc110_spectra, nRows=9, nCols=3,
startRow=3, endRow=9, startColumn=1, endColumn=3), background_spectrum)
TuSC_Spc110_binding <- fretFit(TuSC_Spc110_spectra_bgsub, concentrations=sort(rep(c(1500, 750,
375, 187.5, 93.75, 46.875, 0), 3)), accConc=50)

}
\keyword{datasets}
