% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TuSC_Spc110_spectra}
\alias{TuSC_Spc110_spectra}
\title{Fluorescence spectra of gamma-TuSC-CFP/YFP recorded at a variety of Spc110 concentrations}
\format{29 x 27 matrix.}
\source{
Lyon et al. 2016. Molecular Biology of the Cell 27: 2245, figure 2E (Dimer WT curve). \url{http://www.molbiolcell.org/content/27/14/2245.long}
}
\usage{
TuSC_Spc110_spectra
}
\description{
50 nm gTuSC-CFP/YFP and varying concentration so Spc110 were mixed to allow
assembly of gTuRCs, then fluorescence spectra were recorded on a Molecular Devices SpectraMax
M5 platereader. Columns 1-6 (wells A7-B9) are blank spectra recorded from buffer alone. Columns 7-27 are gTuSC+Spc110 spectra. The Spc110 concentrations are \code{sort(rep(c(1500, 750, 375, 187.5,
93.75, 46.875, 0), 3), decreasing=TRUE)}, in nanomolar. These are raw spectra so background 
spectra from buffer-only controls
must be subtracted before further processing.
}
\keyword{datasets}
